Small fixes

+++ src/collision/grid.cpp
@@ -12,7 +12,7 @@ Grid::Grid(int numCellsSide)
 
 void Grid::buildGrid(ParticleData &particles, BoundingVolume bv, bool parallel)
 {
-    for (int i = 0; i < grid.size(); ++i)
+    for (unsigned int i = 0; i < grid.size(); ++i)
     {
         grid[i].clear();
     }
@@ -24,9 +24,9 @@ void Grid::buildGrid(ParticleData &particles, BoundingVolume bv, bool parallel)
             for (size_t i = r.begin(); i != r.end(); ++i)
             {
                 int x, y, z;
-                x = numCellsSide / bv.length * (particles.pPosition[i].x - bv.corner.x);
-                y = numCellsSide / bv.length * (particles.pPosition[i].y - bv.corner.y);
-                z = numCellsSide / bv.length * (particles.pPosition[i].z - bv.corner.z);
+                x = (int)(numCellsSide / bv.length * (particles.pPosition[i].x - bv.corner.x));
+                y = (int)(numCellsSide / bv.length * (particles.pPosition[i].y - bv.corner.y));
+                z = (int)(numCellsSide / bv.length * (particles.pPosition[i].z - bv.corner.z));
                 if (x < 0)
                     continue;
                 if (y < 0)
@@ -48,12 +48,12 @@ void Grid::buildGrid(ParticleData &particles, BoundingVolume bv, bool parallel)
     }
     else
     {
-        for (int i = 0; i < particles.cardinality; ++i)
+        for (unsigned int i = 0; i < particles.cardinality; ++i)
         {
             int x, y, z;
-            x = numCellsSide / bv.length * (particles.pPosition[i].x - bv.corner.x);
-            y = numCellsSide / bv.length * (particles.pPosition[i].y - bv.corner.y);
-            z = numCellsSide / bv.length * (particles.pPosition[i].z - bv.corner.z);
+            x = (int)(numCellsSide / bv.length * (particles.pPosition[i].x - bv.corner.x));
+            y = (int)(numCellsSide / bv.length * (particles.pPosition[i].y - bv.corner.y));
+            z = (int)(numCellsSide / bv.length * (particles.pPosition[i].z - bv.corner.z));
             if (x < 0)
                 continue;
             if (y < 0)
@@ -89,10 +89,10 @@ void Grid::findCollisions(DistanceConstraintData &constraints, ParticleData &par
 
                 for (int x = -1; x < 2; x++) for (int y = -1; y < 2; y++) for (int z = -1; z < 2; z++)
                 {
-                    for (int j = 0; j < grid[i].size(); ++j)
+                    for (unsigned int j = 0; j < grid[i].size(); ++j)
                     {
                         int xyz = i + x + y * numCellsSide + z * numCellsSide * numCellsSide;
-                        for (int k = i==xyz?j+1:0; k < grid[xyz].size(); ++k)
+                        for (unsigned int k = i==xyz?j+1:0; k < grid[xyz].size(); ++k)
                         {
                             Intersection isect;
                             int idx1 = grid[i][j];
@@ -118,7 +118,7 @@ void Grid::findCollisions(DistanceConstraintData &constraints, ParticleData &par
     }
     else
     {
-        for (int i = 0; i < grid.size(); ++i)
+        for (unsigned int i = 0; i < grid.size(); ++i)
         {
             if (i % numCellsSide == 0)
                 continue;
@@ -131,9 +131,9 @@ void Grid::findCollisions(DistanceConstraintData &constraints, ParticleData &par
                 {
                     for (int z = -1; z < 2; z++)
                     {
-                        for (int j = 0; j < grid[i].size(); ++j) {
+                        for (unsigned int j = 0; j < grid[i].size(); ++j) {
                             int xyz = i + x + y * numCellsSide + z * numCellsSide * numCellsSide;
-                            for (int k = i==xyz?j+1:0; k < grid[xyz].size(); ++k)
+                            for (unsigned int k = i==xyz?j+1:0; k < grid[xyz].size(); ++k)
                             {
                                 Intersection isect;
                                 int idx1 = grid[i][j];
